% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.R
\name{beta_variance}
\alias{beta_variance}
\title{Calculate beta variance from shape parameters}
\usage{
beta_variance(shape1, shape2)
}
\arguments{
\item{shape1}{Numeric. First shape parameter.}

\item{shape2}{Numeric. Second shape parameter.}
}
\value{
Numeric. The variance
}
\description{
Calculate beta variance from shape parameters
}
