% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{betaTrim_mle}
\alias{betaTrim_mle}
\title{Efron null model}
\usage{
betaTrim_mle(x, m, p, pct0 = 0.25, init = c(1, 1, 0.5), iter.max = 200,
  err.max = 1e-06, lb.opt = c(0, 0, 0), ub.opt = c(log(10000), log(10000),
  1))
}
\arguments{
\item{iter.max}{Maximum number of iterations}

\item{lb.opt}{Lower bound}

\item{ub.opt}{Upper bound}
}
\description{
Efron null model
}
\details{
Estimate the null model following Efron's approach.
}
