% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{bsplfun}
\alias{bsplfun}
\title{Create B-spline.}
\usage{
bsplfun(xrange = c(0, 1), breaks = seq(xrange[1], xrange[2], length.out =
  100), k = 4, ncores = 1)
}
\arguments{
\item{xrange}{Support of the spline}

\item{breaks}{Break points of the spline}

\item{k}{Spline order}

\item{ncores}{Number of cores to use}
}
\description{
Create B-spline.
}
\details{
Create a B-spline.
}
