% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{bsplfun.updt}
\alias{bsplfun.updt}
\title{Update B-spline.}
\usage{
bsplfun.updt(ii, bspl)
}
\arguments{
\item{ii}{Index to update}

\item{bspl}{Spline to update}
}
\description{
Update B-spline.
}
\details{
Updates the B-spline at a certain index and returns a new spline.
}
