% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{dbetabinom.vec}
\alias{dbetabinom.vec}
\title{Beta binomial density.}
\usage{
dbetabinom.vec(x, m, rate1, rate2, log = TRUE)
}
\arguments{
\item{x}{Vector of quantiles}

\item{m}{Vector giving numbers of trials}

\item{rate1}{First shape parameter}

\item{rate2}{Second shape parameter}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector describing the density
}
\description{
Beta binomial density.
}
\details{
\code{dbetabinom.vec} returns a vector giving the beta binomial density.
}
