% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{ebBeta}
\alias{ebBeta}
\title{Wrap alternative models.}
\usage{
ebBeta(x, m, p, breaks = seq(0, 1, length.out = 101), k = 4,
  pi.init = rep(1, length(breaks) + k - 4)/(length(breaks) + k - 4),
  pct0 = 0.45, init = NULL, iter.max = 200, err.max = 1e-05,
  ncores = 1, lb.opt = c(0, 0, 0), ub.opt = c(log(10000), log(10000), 1))
}
\arguments{
\item{iter.max}{Max iterations}

\item{ncores}{Number of cores to use}

\item{lb.opt}{Lower bound}

\item{ub.opt}{Upper bound}
}
\description{
Wrap alternative models.
}
\details{
wrapper of EBO and EBE
}
