% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{emBinBspl}
\alias{emBinBspl}
\title{NPBin model.}
\usage{
emBinBspl(x, m, breaks = seq(0, 1, length.out = 101), k = 4,
  pi.init = rep(1, length(breaks) + k - 4)/(length(breaks) + k - 4),
  ncores = 1, err.max = 1e-06, iter.max = 200)
}
\arguments{
\item{x}{x}

\item{m}{m}

\item{breaks}{break points for the spline}

\item{k}{Spline order}

\item{ncores}{Number of cores to use}

\item{err.max}{Max error for the EM algorithm}

\item{iter.max}{Max number of iterations}
}
\value{
A list giving model details
}
\description{
NPBin model.
}
\details{
We remove the basis that only covers bin at the boundary
}
\seealso{
\code{\link{bsplfun}}, \code{\link{getDesignMtx}},
  \code{\link{npbin}}
}
