% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{estNull}
\alias{estNull}
\title{Null model estimation}
\usage{
estNull(x, m, mod, init = NULL, ncores = 1, iter.max = 200,
  err.max = 1e-06, algorithm = "NLOPT_GN_DIRECT_L", pseq = (1:9999)/10000,
  lb = c(0, 0), ub = rep(log(10000), 2))
}
\arguments{
\item{iter.max}{Maximum number of iterations for EM algorithm}

\item{lb}{Lower bounds}

\item{ub}{Upper bounds}

\item{prep}{Info prepared by estNull1}

\item{algoritm}{Numerical library to use}
}
\description{
Null model estimation
}
\details{
estimate the null model
}
