% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{estNull1}
\alias{estNull1}
\title{Null model estimation}
\usage{
estNull1(mod, pseq = (1:9999)/10000, ncores = 1)
}
\arguments{
\item{ncores}{Number of cores to use}
}
\description{
Null model estimation
}
\details{
estimate the null model, part 1
}
