% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{evBspl}
\alias{evBspl}
\title{B-spline evaluation.}
\usage{
evBspl(p, bspl, ncores = 1)
}
\arguments{
\item{p}{p}

\item{bspl}{A B-spline}

\item{ncores}{Number of cores to use}
}
\value{
A matrix giving the evaluation result
}
\description{
B-spline evaluation.
}
\details{
evaluate B-spline function
}
\seealso{
\code{\link{evBsplDrv}}, \code{\link{estNull1}},
  \code{\link{emBspl}}
}
