% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_header_indices}
\alias{get_header_indices}
\title{Get the indices of required fields in a header.}
\usage{
get_header_indices(header)
}
\arguments{
\item{header}{A character vector}
}
\value{
A list mapping required fields to their respective indices
}
\description{
Returns a list providing the index of each required field.
}
\details{
\code{get_header_indices} accepts a character vector, which could be the
entire header in one element separated by whitespace or a vector with
several elements, one for each field. It uses \code{which} to determine
the inidces of the required fields.
}
