% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{getDesignMtx}
\alias{getDesignMtx}
\title{Design matrix extraction}
\usage{
getDesignMtx(bs, x, m, ncores = 1)
}
\arguments{
\item{bs}{A B-spline}

\item{x}{x}

\item{m}{m}

\item{ncores}{Number of cores to use}
}
\description{
Design matrix extraction
}
\details{
get the design matrix c_ij
}
