% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{iBiBsFun}
\alias{iBiBsFun}
\title{Spline basis function integration.}
\usage{
iBiBsFun(bk, x, m, cf)
}
\arguments{
\item{bk}{Pair of points}

\item{x}{x}

\item{m}{m}

\item{cf}{Coefficient}
}
\description{
Spline basis function integration.
}
\details{
Integration over one spline basis function
}
