% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{locfdr2FDR}
\alias{locfdr2FDR}
\title{Convert locfdr to FDR.}
\usage{
locfdr2FDR(locfdr)
}
\arguments{
\item{locfdr}{Local FDR value}
}
\description{
Convert locfdr to FDR.
}
\details{
Convert locfdr to FDR.
}
