% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{logical_to_winning_chip}
\alias{logical_to_winning_chip}
\title{Convert a logical value to a character.}
\usage{
logical_to_winning_chip(logical_val)
}
\arguments{
\item{logical_val}{A logical value}
}
\value{
"P" if input was TRUE, "M" if it was FALSE.
}
\description{
Converts TRUE to "P", FALSE to "M".
}
\details{
This is a utility for construction of the "winning.chip" field required by
NPBin functions.
}
