% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{npbin}
\alias{npbin}
\title{NPBin}
\usage{
npbin(dt.ct, n_breaks, spline_order, pi_init, n_cores)
}
\arguments{
\item{dt.ct}{Data table containing allele counts}

\item{n_breaks}{Number of spline breaks}

\item{spline_order}{Spline order}

\item{pi_init}{Initial weights}

\item{n_cores}{Number of cores to use}
}
\value{
Data table containing model information
}
\description{
NPBin
}
\details{
Perform an NPBin analysis
}
