% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.R
\name{overdispersion}
\alias{overdispersion}
\title{Calculate beta overdispersion from shape parameters}
\usage{
overdispersion(shape1, shape2)
}
\arguments{
\item{shape1}{Numeric. First shape parameter.}

\item{shape2}{Numeric. Second shape parameter.}
}
\value{
Numeric. The overdispersion
}
\description{
Calculate beta overdispersion from shape parameters
}
