% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimated_null.R
\name{plot_estimated_null}
\alias{plot_estimated_null}
\title{Plot an estimated null distribution}
\usage{
plot_estimated_null(p_hat, shape1, shape2)
}
\arguments{
\item{p_hat}{Numeric. Vector of observed ref allele proportions.}

\item{shape1}{Numeric. First shape parameter of the null distribution.}

\item{shape2}{Numeric. Second shape parameter of the null distribution.}
}
\description{
Plot an estimated null distribution
}
