% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{rank2nhit}
\alias{rank2nhit}
\title{Convert ranking to the number of discoveries}
\usage{
rank2nhit(r, id)
}
\arguments{
\item{r}{Vector of ranks}

\item{id}{id}
}
\value{
vector indicating discoveries
}
\description{
Convert ranking to the number of discoveries
}
\details{
Convert ranking to the number of discoveries
}
