% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.R
\name{symmetric_beta_shape_from_variance}
\alias{symmetric_beta_shape_from_variance}
\title{Calculate a symmetric shape parameter from the variance}
\usage{
symmetric_beta_shape_from_variance(variance)
}
\arguments{
\item{variance}{Numeric. The variance}
}
\value{
Numeric. The shape parameter
}
\description{
Calculate a symmetric shape parameter from the variance
}
