% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{emBspl}
\alias{emBspl}
\title{Direct B-spline estimation.}
\usage{
emBspl(x, m, p, breaks = seq(0, 1, length.out = 101), k = 4,
  pi.init = rep(1, length(breaks) + k - 4)/(length(breaks) + k - 4),
  ncores = 1, err.max = 1e-05, iter.max = 200)
}
\arguments{
\item{k}{Spline order}
}
\description{
Direct B-spline estimation.
}
\details{
Estimate the overall model directly via B-spline. Remove the basis that only
covers bin at the boundary.
}
