% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{evBsplDrv}
\alias{evBsplDrv}
\title{B-spline derivative evaluation.}
\usage{
evBsplDrv(p, bspl, ncores = 1)
}
\arguments{
\item{p}{p}

\item{bspl}{A B-spline}

\item{ncores}{Number of cores to use}
}
\value{
A matrix giving the evaluation result
}
\description{
B-spline derivative evaluation.
}
\details{
evaluate the derivative of B-spline
}
\seealso{
\code{\link{evBspl}}, \code{\link{estNull1}}
}
