% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_weights.R
\name{initialize_weights}
\alias{initialize_weights}
\title{Initialize weights}
\usage{
initialize_weights(data_table, n_breaks, spline_order)
}
\arguments{
\item{data_table}{Input data table}

\item{n_breaks}{Number of breaks for the histogram}

\item{spline_order}{Order of splines}
}
\value{
Histogram values
}
\description{
Initialize weights
}
\details{
Initialize the weights using the histogram of p_hat
}
\seealso{
\code{\link{npbin}}, \code{\link{emBspl}}, \code{\link{ebBeta}}
}
