% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbin.R
\name{intBinBspl}
\alias{intBinBspl}
\title{Binomial and B-spline integration.}
\usage{
intBinBspl(bs, x, m)
}
\arguments{
\item{bs}{A B-spline}

\item{x}{x}

\item{m}{m}
}
\value{
The integral value.
}
\description{
Binomial and B-spline integration.
}
\details{
Integration of Binomial + B spline.
}
