% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimated_null.R
\name{plot_estimated_null}
\alias{plot_estimated_null}
\title{Plot an estimated null distribution}
\usage{
plot_estimated_null(p_hat, shape1_shape2, shape3_shape4 = NULL)
}
\arguments{
\item{p_hat}{Numeric. Vector of observed ref allele proportions.}

\item{shape1_shape2}{Numeric. Shape parameters of the first null
distribution.}

\item{shape3_shape4}{Numeric. Shape parameters of the second null
distribution.}
}
\description{
optionally superimpose two null distributions
}
